move syntax_highlight doc filter in compatibility module and define it conditionally
